fix: define types
