+++ src/pages/Detailed-item/Detailed-item.tsx
@@ -6,6 +6,7 @@ import cn from 'classnames';
 import {ThemeContext} from '../../services/theme/ThemeProvider';
 import {useDataFetching} from '../../hooks/useDataFetching';
 
+import {AnimeData} from '../../types/animeData';
 import styles from './Detailed-item.module.css';
 
 function DetailedItem() {
@@ -24,8 +25,9 @@ function DetailedItem() {
 
     const params = useParams();
 
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const anime: any = useDataFetching(`https://anime-db.p.rapidapi.com/anime/by-id/${params.id}`);
+    const anime: AnimeData | null = useDataFetching(
+        `https://anime-db.p.rapidapi.com/anime/by-id/${params.id}`
+    );
 
     const handleClick = () => {
         navigate('/anime-list');
